feat: complete beauty-refactor of `blext build` w/ugly code
Architecture of `blext.pydeps.network` was altered to use a `ThreadPoolExecutor`. Both this and `blext.pack` now support callbacks instead of `rich` calls, causing a lot of simplifications to this internal code. However, the awesome new UI implemented in `build` makes for incredibly verbose code, which **absolutely must** be refactored away (it does work now, but there is no reuse). Moreover, several bugs were noticed: There's no reason to reject building extensions for just a few platforms; the download-prepack-pack chain is not robust at detecting when a step needs doing; so on.

Soundly closes #16. Progress on #17. Closes #6. Should implement #12, since #13 would be easy to implement while we're here.