feat: new ui for pre-packing feat. many fixes
With better UX comes easier code to implement features in. Vendoring is now implemented with `uv.lock`. Some defaults have been shifted, including a default zip location for script extensions being next to the script itself. Detecting when to download and/or pre-pack is now robust; only wheel hash-checks are missing for it to be theoretically sound. Finally, everything is now ready for arbitrary pre-packing.

BREAKING CHANGE: All is ready for #13, and implementing #12 is now very doable. Closes #20.