⚡️ perf: implemented `lru_method` to memoize `BLExtInfo` w/o memory leaks
Since many objects of interest are immutable, it's valid to use `lru_cache` decorators. However, this normally results in a memory leak, as instances will now never GC, due to the references held by the decorated methods. By creating a custom decorator, a weak reference can be used instead, allowing for an elegant method of optimizing performance of objects like `BLExtInfo` (which can't keep a `GlobalCache` itself, but which will generally be called with the same one throughout a run).