✨ feat!: implemented smart build-by-platform as default
The default is now to build extension `.zip`s into `build/` of the project room. To this end, `--output` now supports a few macros like `%platform`, which are optional/required depending on the context. Platforms are overall now "smooshable" courtesy the very similar `BLPlatforms` (as opposed to `BLPlatform`) object, and a number of change incl. to version string handling, extension filename generation, and error-propagation through the wheel selection, are now implemented. Finally, a quick fix to ensure the cache directory for `blext` is created on startup.