🐛 fix: greatly improved version strings by vacuuming `BLVersionSource`
Completely removing `BLVersionSource` was the right move, as its functionality should really always have been encapsulated by `BLVersion`. The one component that we may want to bring back one day is the `git` based sourcing, but that should be done as a `BLRelease` that generates `BLVersion` aka. `BLReleaseGit` or some such. A final comment is that the "smooshing" logic between `BLVersion`s was streamlined to be more optimistic about things, in order to separate the filtering and smooshing stages for greater flexibility.